+++ app.js
@@ -10,6 +10,9 @@ var	http			=	require('http');
 var	io				=	require('socket.io');
 var	connect			=	require('connect');
 var SM 				=	require('./server/modules/socket-manager');
+var CM 				=	require('./server/modules/category-manager');
+var	jsdom			=	require('jsdom');
+var	func			=	require('./server/controllers/func.js');
 
 var app				=	express();
 
@@ -22,7 +25,9 @@ server.listen(3000);
 var	socket			=	io.listen(server);
 require('./server/router')(app, socket);
 
-
+/*func.getTitleFromURL('http://xforum.slashmanx.com', function(title) {
+	console.log(title)
+})*/
 socket.set('authorization', function (data, accept) {
     // check if there's a cookie header
     if (data.headers.cookie) {
@@ -46,13 +51,24 @@ socket.sockets.on('connection', function(client){
 	client.join(hs.sessionID);
 
 	client.on('leavingTopic', function(data){
-		/* Insert function to access the express_session table for hs.sessionID to get user ID */
 		SM.getUserIDFromSession(hs.sessionID.slice(2), function(e, a) {
 			if(a) {
 				SM.setLastReadTime(JSON.parse(a.session).user._id, data.topic, function(e, o) {
 				});
 			}
 		});
+	})
+	
+	client.on('getTitleFromURL', function(url, callback) {
+			func.getTitleFromURL(url, function(title) {
+				callback(title);
+			})
+	})
+	
+	client.on('getCategoryDetails', function(data, callback) {
+		CM.getDetails(data.id, function(e, cat) {
+			callback(cat);
+		})
 	})
 	client.on('disconnect', function(){
 		client.leave(hs.sessionID);